refactor: split profile into components and other code refactors
